% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonvt-package.R
\docType{package}
\name{geojsonvt-package}
\alias{geojsonvt}
\alias{geojsonvt-package}
\title{Provides access to Mapbox Geojson-vt-cpp library}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\examples{
\dontrun{

# Include a dependency to geojsonvt inside a cpp file
// [[Rcpp::depends(geojsonvt)]]
include <Rcpp.h>
}
}
