% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonvt-package.R
\docType{package}
\name{geojsonvt-package}
\alias{geojsonvt}
\alias{geojsonvt-package}
\title{Provides access to Mapbox Geojson-vt-cpp library}
\description{
Converts geojson into vector-tile format to be used inside 'Mapbox' maps
}
\examples{
\dontrun{

# Include a dependency to geojsonvt inside a cpp file
// [[Rcpp::depends(geojsonvt)]]
include <Rcpp.h>
}
}
\author{
\strong{Maintainer}: David Cooley \email{dcooley@symbolix.com.au}

Other contributors:
\itemize{
  \item Mapbox (Author of c++ geojson-vt-cpp library) [contributor]
}

}
